refactor: architecture

+++ main-service/src/main/java/ru/tokmakov/dto/complation/UpdateCompilationRequest.java
@@ -13,7 +13,7 @@ import java.util.Set;
 @AllArgsConstructor
 @NoArgsConstructor
 public class UpdateCompilationRequest {
-    private Set<Integer> events;
+    private Set<Long> events;
     private Boolean pinned;
     @Size(min = 1, max = 50)
     private String title;
